requirements.txt
streamlit>=1.32,<2
psycopg2-binary==2.9.9
pandas>=2.2,<3
python-dateutil>=2.9
python-dotenv>=1.0
Unidecode>=1.3